+++ trunk/mos/net/routing/dv.c
@@ -66,7 +66,7 @@ static bool isForwardTimerActive(void)
 
 static inline bool isRoutingInfoValid(void)
 {
-    return timeAfter32(lastRootMessageTime + ROUTING_INFO_VALID_TIME, getJiffies());
+    return timeAfter32(lastRootMessageTime + ROUTING_INFO_VALID_TIME, (uint32_t)getJiffies());
 }
 
 void initRouting(void)
